Moved registerform out of subfolder of components
