+++ src/components/RegisterForm.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { useAppDispatch, useAppSelector } from '../../hooks/redux.ts'
+import { useAppDispatch, useAppSelector } from '../hooks/redux'
 
 function RegisterForm() {
   const dispatch = useAppDispatch()
