+++ src/pages/Register/index.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom'
-import RegisterForm from '../../components/RegisterForm/RegisterForm.tsx'
+import RegisterForm from '../../components/RegisterForm'
 
 function RegisterPage() {
   const navigate = useNavigate()
